Change SearchBar into a component
