+++ frontend/src/jss/AppStyle.js
@@ -1,11 +1,11 @@
 const AppStyle = () => {
     return {
         root: {
-            // height: 768,
-            // width: 1024,
-            // maxHeight: 768,
-            // maxWidth: 1024
-            minWidth: 750
+            height: 768,
+            width: 1024,
+            maxHeight: 768,
+            maxWidth: 1024,
+            position: 'relative'
         }
     };
 };
