+++ frontend/src/jss/components/SearchBarStyle.js
@@ -1,0 +1,15 @@
+const SearchBarStyle = () => {
+    return {
+        searchBar: {
+            display: "flex",
+            width: "100%",
+            paddingBottom: "10px",
+            flexWrap: 'wrap'
+        },
+        textField: {
+            width: "110px",
+        },
+    };
+};
+
+export default SearchBarStyle;
